Refactor, handle err, simplify
